backend/requirements/base.txt
Django>=6.0,<6.1
djangorestframework>=3.16,<3.17
psycopg[binary]>=3.2,<3.4
python-dotenv>=1.0,<2.0